Changed builders to static inner classes
