+++ src/test/java/r11/citrus/s3/S3AbstractHost.java
@@ -1,4 +1,4 @@
-package r11.citrus.s3.host;
+package r11.citrus.s3;
 
 import org.gaul.s3proxy.S3Proxy;
 import org.gaul.shaded.org.eclipse.jetty.util.component.AbstractLifeCycle;
@@ -23,7 +23,7 @@ public class S3AbstractHost {
      * Constructor Uses builder configuration to build an object
      * @param builder
      */
-    public S3AbstractHost(S3AbstractHostBuilder builder){
+    public S3AbstractHost(S3AbstractHost.Builder builder){
         Properties properties = new Properties();
         properties.setProperty("jclouds.filesystem.basedir", builder.getLocalBlobPath());
         properties.setProperty("jclouds.regions", builder.getRegion());
@@ -62,8 +62,8 @@ public class S3AbstractHost {
      * Static method for obtaining the builder
      * @return
      */
-    public static S3AbstractHostBuilder builder(){
-        return new S3AbstractHostBuilder();
+    public static S3AbstractHost.Builder builder(){
+        return new S3AbstractHost.Builder();
     }
 
     /**
@@ -142,4 +142,158 @@ public class S3AbstractHost {
         this.blobStore.removeBlob(bucket, key);
     }
 
+    /**
+     * The instance of this class serves as builder to S3AbstractHost. It gathers all required
+     * data to initialize an instance of the class. Operation build returns an object.
+     */
+    static class Builder {
+
+        /**
+         * initialized with default value
+         */
+        private String service = "filesystem";
+        /**
+         * initialized with default value
+         */
+        private String region = "eu-east-1";
+        /**
+         * initialized with default value
+         */
+        private String localBlobPath = "./tmp/blobstore";
+        /**
+         * initialized with default value
+         */
+        private String host = "http://127.0.0.1:4566";
+        private String accessKey;
+        private String secretKey;
+
+        /**
+         * service getter
+         *
+         * @return
+         */
+        public String getService() {
+            return service;
+        }
+
+        /**
+         * region getter
+         *
+         * @return
+         */
+        public String getRegion() {
+            return region;
+        }
+
+        /**
+         * localBlobPath getter
+         *
+         * @return
+         */
+        public String getLocalBlobPath() {
+            return localBlobPath;
+        }
+
+        /**
+         * host getter
+         *
+         * @return
+         */
+        public String getHost() {
+            return host;
+        }
+
+        /**
+         * accessKey getter
+         *
+         * @return
+         */
+        public String getAccessKey() {
+            return accessKey;
+        }
+
+        /**
+         * secretKey getter
+         *
+         * @return
+         */
+        public String getSecretKey() {
+            return secretKey;
+        }
+
+        /**
+         * Sets value during building
+         *
+         * @param service
+         * @return
+         */
+        public S3AbstractHost.Builder service(String service) {
+            this.service = service;
+            return this;
+        }
+
+        /**
+         * Sets value during building
+         *
+         * @param region
+         * @return
+         */
+        public S3AbstractHost.Builder region(String region) {
+            this.region = region;
+            return this;
+        }
+
+        /**
+         * Sets value during building
+         *
+         * @param localBlobPath
+         * @return
+         */
+        public S3AbstractHost.Builder localBlobPath(String localBlobPath) {
+            this.localBlobPath = localBlobPath;
+            return this;
+        }
+
+        /**
+         * Sets value during building
+         *
+         * @param host
+         * @return
+         */
+        public S3AbstractHost.Builder host(String host) {
+            this.host = host;
+            return this;
+        }
+
+        /**
+         * Sets value during building
+         *
+         * @param accessKey
+         * @return
+         */
+        public S3AbstractHost.Builder accessKey(String accessKey) {
+            this.accessKey = accessKey;
+            return this;
+        }
+
+        /**
+         * Sets value during building
+         *
+         * @param secretKey
+         * @return
+         */
+        public S3AbstractHost.Builder secretKey(String secretKey) {
+            this.secretKey = secretKey;
+            return this;
+        }
+
+        /**
+         * Creates S3AbstractHost instance based on provided data.
+         *
+         * @return
+         */
+        public S3AbstractHost build() {
+            return new S3AbstractHost(this);
+        }
+    }
 }
